reorg BottomBar menu item classes
